Final Packing

+++ app/logics/renderTools/Point3d.java
@@ -5,16 +5,16 @@ package logics.renderTools;
  */
 public class Point3d {
 
-    private double x;
-    private double y;
+    private float x;
+    private float y;
 
-    public Point3d(double x, double y, double z) {
+    public Point3d(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    private double z;
+    private float z;
 
     @Override
     public String toString() {
@@ -25,27 +25,27 @@ public class Point3d {
                 '}';
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public double getZ() {
+    public float getZ() {
         return z;
     }
 
-    public void setZ(double z) {
+    public void setZ(float z) {
         this.z = z;
     }
 }
